Move setDefaultOrCLocale function to libdevcore
Move setDefaultOrCLocale helper function to libdevcore to reduce code duplication since it's defined in many main.cpp files. Also move the call which changes the code page on Windows to this function so that it's invoked in all Aleth binaries.
