+++ aleth/main.cpp
@@ -95,29 +95,6 @@ bool isFalse(std::string const& _m)
     return _m == "off" || _m == "no" || _m == "false" || _m == "0";
 }
 
-/*
-The equivalent of setlocale(LC_ALL, “C”) is called before any user code is run.
-If the user has an invalid environment setting then it is possible for the call
-to set locale to fail, so there are only two possible actions, the first is to
-throw a runtime exception and cause the program to quit (default behaviour),
-or the second is to modify the environment to something sensible (least
-surprising behaviour).
-
-The follow code produces the least surprising behaviour. It will use the user
-specified default locale if it is valid, and if not then it will modify the
-environment the process is running in to use a sensible default. This also means
-that users do not need to install language packs for their OS.
-*/
-void setDefaultOrCLocale()
-{
-#if __unix__
-    if (!std::setlocale(LC_ALL, ""))
-    {
-        setenv("LC_ALL", "C", 1);
-    }
-#endif
-}
-
 void importPresale(KeyManager& _km, string const& _file, function<string()> _pass)
 {
     KeyPair k = _km.presaleSecret(contentsString(_file), [&](bool){ return _pass(); });
@@ -179,12 +156,6 @@ bool ExitHandler::s_shouldExit = false;
 int main(int argc, char** argv)
 {
     setDefaultOrCLocale();
-
-#if defined(_WIN32)
-    // Need to change the code page from the default OEM code page (437) so that
-    // UTF-8 characters are displayed correctly in the console
-    SetConsoleOutputCP(CP_UTF8);
-#endif
 
     // Init secp256k1 context by calling one of the functions.
     toPublic({});
