+++ libdevcore/Common.cpp
@@ -19,6 +19,10 @@
 #include "Exceptions.h"
 #include "Log.h"
 
+#if defined(_WIN32)
+#include <windows.h>
+#endif
+
 #include <aleth/buildinfo.h>
 
 using namespace std;
@@ -93,4 +97,32 @@ string inUnits(bigint const& _b, strings const& _units)
     return ret.str();
 }
 
+/*
+The equivalent of setlocale(LC_ALL, “C”) is called before any user code is run.
+If the user has an invalid environment setting then it is possible for the call
+to set locale to fail, so there are only two possible actions, the first is to
+throw a runtime exception and cause the program to quit (default behaviour),
+or the second is to modify the environment to something sensible (least
+surprising behaviour).
+
+The follow code produces the least surprising behaviour. It will use the user
+specified default locale if it is valid, and if not then it will modify the
+environment the process is running in to use a sensible default. This also means
+that users do not need to install language packs for their OS.
+*/
+void setDefaultOrCLocale()
+{
+#if __unix__
+    if (!std::setlocale(LC_ALL, ""))
+    {
+        setenv("LC_ALL", "C", 1);
+    }
+#endif
+
+#if defined(_WIN32)
+    // Change the code page from the default OEM code page (437) so that UTF-8 characters are
+    // displayed correctly in the console.
+    SetConsoleOutputCP(CP_UTF8);
+#endif
+}
 }
