+++ libdevcore/Common.h
@@ -302,4 +302,5 @@ enum class WithExisting: int
 /// Get the current time in seconds since the epoch in UTC
 int64_t utcTime();
 
+void setDefaultOrCLocale();
 }
